+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - TextMesh Pro package dependency
 - InputSystem.OnDeviceInputChanged: event fired when any input from any connected device changed
+- SpriteDisplayer: replaces any occurrences of {inputName} by the corresponding TextMeshPro Sprite tag.
 
 ### Changed
 - Update Input System package to 1.2.0